requirements.txt
annotated-types==0.6.0
anyio==4.2.0
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai==1.6.1
openai-whisper==20231117
pycryptodomex==3.19.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
sympy==1.12
tiktoken==0.5.2
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
websockets==12.0
yt-dlp
